Add service layer, BaseEntity, etc
 - add simple CarService service layer
 - introduce BaseEntity to hold common id field
 - change amount type to float
 - change from in-memory DB to actual db
 - add simple test for CarController

+++ src/main/resources/schema.sql
@@ -15,7 +15,7 @@ CREATE INDEX cars_make ON cars (make);
 CREATE TABLE prices
 (
     id     INTEGER IDENTITY PRIMARY KEY,
-    amount BIGINT  NOT NULL,
+    amount FLOAT   NOT NULL,
     car_id INTEGER NOT NULL
 
 );
